Add OverviewPage class with functionality for toggling planet details and buff bar
