+++ src/ctxpage/conf-options.js
@@ -87,6 +87,8 @@ const _options = {
     standardFleet: false,
     standardFleetId: 0,
   },
+  overview_display_planet_details: true,
+  overview_display_planet_buffBar: true,
 };
 
 export function initConfOptions(options) {
